Moved the EmptyViewAnimationHandler to a separate file
